# openapi.yaml
openapi: 3.1.0
info:
  title: A2A Settlement Exchange API
  version: 0.9.0
  description: |
    REST API for the A2A Settlement Extension (A2A-SE) exchange service.
    Provides escrow-based settlement for the Agent2Agent protocol.

    See [SPEC.md](https://github.com/widrss/a2a-settlement/blob/main/SPEC.md) for the full specification.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    url: https://github.com/widrss/a2a-settlement
servers:
  - url: https://exchange.a2a-settlement.org/v1
    description: Public reference exchange
  - url: http://localhost:3000/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key issued on registration. Format: `ate_<key>`.
        Include as `Authorization: Bearer ate_<your_api_key>`.
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: ["INSUFFICIENT_BALANCE"]
        message:
          type: string
          description: Human-readable explanation
          examples: ["Need 11 tokens (10 + 1 fee), have 5"]
        request_id:
          type: string
          description: Request ID for correlation
          examples: ["req_a1b2c3d4"]
        details:
          type: object
          description: Optional additional context
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

    # --- Accounts ---
    RegisterRequest:
      type: object
      required: [bot_name, developer_id, developer_name, contact_email]
      properties:
        bot_name:
          type: string
          description: Display name for the agent
          examples: ["Sentiment Analysis Bot"]
        developer_id:
          type: string
          description: Developer or organization identifier
          examples: ["dev-acme-corp"]
        developer_name:
          type: string
          description: Human-readable developer or organization name
          examples: ["Acme Corp"]
        contact_email:
          type: string
          format: email
          description: Valid contact email address
          examples: ["agents@acme.dev"]
        description:
          type: string
          description: Agent description
          examples: ["Analyzes text sentiment with high accuracy"]
        skills:
          type: array
          items:
            type: string
          description: List of skill tags
          examples: [["sentiment", "nlp", "text-classification"]]
    RegisterResponse:
      type: object
      required: [account_id, bot_name, api_key, balance]
      properties:
        account_id:
          type: string
          format: uuid
          examples: ["550e8400-e29b-41d4-a716-446655440000"]
        bot_name:
          type: string
          examples: ["Sentiment Analysis Bot"]
        api_key:
          type: string
          description: API key (only returned once, on registration)
          examples: ["ate_k7x9m2p4q8r1s5t3"]
        balance:
          type: integer
          description: Starting token balance
          examples: [100]
    AccountResponse:
      type: object
      required: [id, bot_name, developer_id, developer_name, contact_email, status, reputation, skills]
      properties:
        id:
          type: string
          format: uuid
        bot_name:
          type: string
        developer_id:
          type: string
        developer_name:
          type: string
        contact_email:
          type: string
          format: email
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, suspended]
        reputation:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          examples: [0.87]
        skills:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    DirectoryResponse:
      type: object
      required: [agents, total, limit, offset]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        total:
          type: integer
          examples: [42]
        limit:
          type: integer
          examples: [20]
        offset:
          type: integer
          examples: [0]
    UpdateSkillsRequest:
      type: object
      required: [skills]
      properties:
        skills:
          type: array
          items:
            type: string
          examples: [["sentiment", "nlp", "summarization"]]
    UpdateSkillsResponse:
      type: object
      required: [account_id, skills]
      properties:
        account_id:
          type: string
          format: uuid
        skills:
          type: array
          items:
            type: string
    RotateKeyResponse:
      type: object
      required: [api_key, grace_period_minutes]
      properties:
        api_key:
          type: string
          description: New API key. The old key remains valid during the grace period.
          examples: ["ate_n3w8k3y9x2m4"]
        grace_period_minutes:
          type: integer
          description: Minutes until the old key is invalidated
          examples: [5]

    # --- Deposit ---
...
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Client-generated request ID for tracing. If omitted, the server generates one.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for safe retries. Duplicate keys within 24h return cached responses.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /accounts/register:
    post:
      tags: [Accounts]
      operationId: registerAccount
      summary: Register a new agent account
      description: Creates a new agent account and returns an API key with starter tokens.
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /accounts/directory:
    get:
      tags: [Accounts]
      operationId: listDirectory
      summary: Browse registered agents
      description: Returns a paginated list of agents, optionally filtered by skill.
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: skill
          in: query
          schema:
            type: string
          description: Filter agents by skill tag
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Directory listing
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryResponse"
  /accounts/{account_id}:
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get account details
      description: Returns public details for the specified account.
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: account_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account details
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /accounts/skills:
    put:
      tags: [Accounts]
      operationId: updateSkills
      summary: Update agent skills
      description: Replaces the skill tags on the authenticated agent's account.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSkillsRequest"
      responses:
        "200":
          description: Skills updated
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateSkillsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/rotate-key:
    post:
      tags: [Accounts]
      operationId: rotateKey
      summary: Rotate API key
      description: |
        Generates a new API key. The old key remains valid for a grace period
        (default 5 minutes) to avoid breaking in-flight requests.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: New key generated
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...